+++ src/ui/components/Icon/index.tsx
@@ -36,7 +36,8 @@ export const svgRegistry = {
   'circle-check': '/images/icons/circle-check.svg',
   pencil: '/images/icons/pencil.svg',
   'circle-info': '/images/icons/circle-info.svg',
-  bitcoin: './images/icons/bitcoin.svg'
+  bitcoin: './images/icons/bitcoin.svg',
+  'circle-question': '/images/icons/circle-question.svg'
 };
 
 const iconImgList: Array<IconTypes> = ['success', 'delete', 'btc'];